requirements.txt
flask
pandas
openpyxl
pyreadstat
python-docx
matplotlib
seaborn
scipy
statsmodels
pingouin
